Revive XDM rpm
- Update to version 1.1.14

- Remove unnecessary patches for crypt.h and libdl
The package was using the version 1.1.11 which requires these patches.
The lastest version has it already included making the patches no longer
needed.

- New patch for version 1.1.14 for audit and systemd unit file
Since version 1.1.11 the files had changed and the patches no longer
worked for the newest version (1.1.14). The new patches are basically
the same as before, only for version 1.1.14

- Add support to systemd notify
The commit https://gitlab.freedesktop.org/xorg/app/xdm/-/commit/8eee942be0db913511b4c53f94829e1201dfdf00
added (optional) support to systemd startup notification. I've added the
systemd-devel as a build requirement making it default.

- Add patch to fix build when auto-identifying the libsystemd
There is a bug where the `--with-systemd-daemon` option is always being
evaluated to `true` when it auto-identify if the libsystemd is present in
the system. With this patch, it fixes that issue. Also, there is a Merge
Request fixing the bug, so it is most likely that the patch will be
removed soon.

- Add patch to set -nolisten tcp
It is basically a copy from the patch from the xorg-xdm package for Arch
Linux.